+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Bug fixes:
 * Guard against unterminated ranges in file matching patterns (#2556, @aardvark179).
 * Fixed `rb_proc_new` to return a proc that will pass all required arguments to C (#2556, @aardvark179).
 * Fixed `String#split` to return empty array when splitting all whitespace on whitespace (#2565, @bjfish).
+* Raise `RangeError` for `Time.at(bignum)` (#2580, @eregon).
+* Fix `Integer#{<<,>>}` with RHS bignum and long (@eregon).
 
 Compatibility: